is_logined func
