+++ .gitignore
@@ -1,1 +1,2 @@
 database.py
+*.pyc
